requirements.txt
# 模型相关
transformers>=4.36.0
bitsandbytes>=0.41.0
# PyTorch 2.6.0及以上版本，支持CUDA 12.1/12.4
torch>=2.6.0
# 与PyTorch 2.6.0兼容的版本
# 注意：安装时需根据CUDA版本选择合适的包，例如使用CUDA 12.4时：
# pip install torch==2.6.0+cu124 torchvision==0.21.0+cu124 torchaudio==2.6.0+cu124 --index-url https://download.pytorch.org/whl/cu124
torchvision>=0.21.0
torchaudio>=2.6.0
accelerate>=0.24.0

# Web框架
fastapi>=0.104.0
uvicorn>=0.24.0
transformers-stream-generator>=0.0.5
pydantic>=2.0.0

# 数据获取与处理
akshare>=1.10.0
baostock>=0.8.8
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0

# 缓存系统
cachetools>=5.3.0
diskcache>=5.6.0

# HTTP请求与网络
# 推荐使用charset_normalizer替代chardet（更现代的字符编码检测库）
# 注意：urllib3 2.6.x与requests可能存在版本冲突，建议使用兼容版本
urllib3>=1.26.16,<2.7.0
requests>=2.31.0,<3.0.0
# chardet>=5.2.0  # 可选，建议使用charset_normalizer替代
charset_normalizer>=3.3.0
aiohttp>=3.9.0

# 网页解析
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 新闻与搜索
# 可选：Serper.dev API客户端
# serperdev>=1.0.0

# 报告生成
python-docx>=1.1.0
markdown>=3.5.0

# 工具类
python-dotenv>=1.0.0
python-multipart>=0.0.6

# 数据库
pysqlite3>=0.5.0

# 并发与异步
# asyncio是Python标准库，无需单独安装
# asyncio>=3.4.3

# 文本处理
regex>=2023.10.3

# 数据可视化（可选，用于生成报告图表）
matplotlib>=3.8.0
seaborn>=0.13.0

# 新增依赖
# 股票数据备用数据源
tickflow>=0.1.0  # 备用数据源

# 网络安全
pyopenssl>=23.0.0

# 系统工具
psutil>=5.9.0  # 系统资源监控

# 日志
loguru>=0.7.0  # 更强大的日志库

